+++ src/main/java/Controller/Music.java
@@ -1,5 +1,6 @@
 package Controller;
 
+import java.io.File;
 import java.io.IOException;
 import java.util.List;
 
@@ -8,6 +9,9 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+
+import com.oreilly.servlet.MultipartRequest;
+import com.oreilly.servlet.multipart.DefaultFileRenamePolicy;
 
 import DAO.MusicDAO;
 import DTO.MusicDTO;
@@ -27,6 +31,7 @@ public class Music extends HttpServlet {
 				List<MusicDTO> musicChartList = dao.musicChartList();
 				request.setAttribute("musicChartList", musicChartList);
 				request.getRequestDispatcher("/index.jsp").forward(request, response);
+				
 			} else if (uri.equals("/allList.music")) {
 				MusicDAO dao = MusicDAO.getInstance();
 				List<MusicDTO> musicList = dao.musicAllList();
@@ -34,6 +39,57 @@ public class Music extends HttpServlet {
 				request.setAttribute("list", musicList);
 				request.setAttribute("nickname", nickName);
 				request.getRequestDispatcher("/admin/adminMusic.jsp").forward(request, response);
+				
+			}else if(uri.equals("/adminMusic.music")) {
+				String nickName = request.getParameter("nickname");
+				request.setAttribute("nickname", nickName);
+				request.getRequestDispatcher("/admin/adminMusic.jsp").forward(request, response);
+				
+			}else if(uri.equals("/addMusic.music")) {
+				// 파일등록
+				int maxSize = 1024*1024*10;
+				String savePath = request.getServletContext().getRealPath("/image/albumImg");
+				File fileSavePath = new File(savePath);
+
+				if(!fileSavePath.exists()){
+					fileSavePath.mkdir(); 
+				}
+
+				MultipartRequest multi = new MultipartRequest(request,savePath,maxSize,"UTF8",new DefaultFileRenamePolicy());
+
+				int getSeq = MusicDAO.getInstance().getSeq();
+
+				String musicName = multi.getParameter("musicName");
+				String musicArtist = multi.getParameter("musicArtist");
+				String musicAlbum = multi.getParameter("musicAlbum");
+				String musicGenre = multi.getParameter("musicGenre");
+				int musicChart = Integer.parseInt(multi.getParameter("musicChart"));
+				String musicLylics = multi.getParameter("musicLylics");
+				MusicDAO dao = MusicDAO.getInstance();
+
+				String imgSysName = multi.getFilesystemName("imgFile");
+				String mp3SysName = multi.getFilesystemName("mp3File");
+
+				MusicDTO dto = new MusicDTO(getSeq,musicName,musicArtist,musicAlbum,imgSysName,mp3SysName,musicChart,musicGenre,musicLylics);
+				dao.addMusic(dto);
+				response.sendRedirect("/allList.music");
+
+			}else if(uri.equals("/deleteMusic.music")) {
+				int seq = Integer.parseInt(request.getParameter("seq"));
+				MusicDAO dao = MusicDAO.getInstance();
+				dao.deleteMusic(seq);
+				response.sendRedirect("/allList.music");
+				
+			}else if(uri.equals("/update.music")) {
+				int seq = Integer.parseInt(request.getParameter("modifySeq"));
+				String musicName = request.getParameter("updateName");
+				String musicArtist = request.getParameter("updateArtist");
+				String musicAlbum = request.getParameter("updateAlbum");
+				String musicGenre = request.getParameter("updateGenre");
+				
+				MusicDAO dao = MusicDAO.getInstance();
+				dao.updateMusic(musicName, musicArtist, musicAlbum, musicGenre, seq);
+				response.sendRedirect("/allList.music");
 			}
 		} catch (Exception e) {
 			e.printStackTrace();
